Add image bundler
The image bundler takes a YAML config file of images to download if
updated, and if there are updates it bundles them into a tarball (or
tar.gz if compress is set in the YAML).

See the YAML file for documentation on how to configure it. All values
are requied